fix(manage): move withdrawal questions into representation questions in shared lib
